requirements.txt
streamlit
torch
Pillow